% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SECEX12FOB12wb.R
\docType{data}
\name{codpaises.SECEX12FOB12}
\alias{codpaises.SECEX12FOB12}
\title{Codigos dos paises - SECEX MDIC}
\format{Banco de dados com 278 observacao(oes) e 3 variavel(is):
\describe{
  \item{"CO_PAIS"}{Codigo dos pais}
  \item{"NO_PAIS_ING"}{Nome dos pais (ingles)}
  \item{"TERRIT_PAISES"}{Codigo de territorio (Numerico)}
}}
\source{
\url{http://www.mdic.gov.br/arquivos/dwnl_1362158650.pdf}
}
\usage{
codpaises.SECEX12FOB12
}
\description{
Conjunto de dados contendo os codigos dos paises utilizado pela SECEX/MDIC.
}
\keyword{datasets}
