\name{codpaisesSECEX12FOB12}

\alias{codpaisesSECEX12FOB12}

\docType{data}

\title{
Codigo dos paises utilizado pela SECEX.
}

\description{
Este conjunto de dados contem o codigo dos paises utilizado pela SECEX/MDIC.
}
\usage{data("codpaisesSECEX12FOB12")}
\format{
  Uma tabela com 252 observacoes com a seguinte variavel.
  \describe{
    \item{\code{cod_paises}}{codigo do pais}
  }
}

\source{A tabela pode ser encontrada na integra em
\link{http://www.mdic.gov.br/arquivos/dwnl_1362158650.pdf
}
