% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONFAZ12WB.R
\docType{data}
\name{codterritCONFAZ12}
\alias{codterritCONFAZ12}
\title{Codigo dos territorios - CONFAZ}
\format{um data frame com 86 observacao(oes) e 3 variavel(is):
\describe{
  \item{conta}{Tipo de variavel}
  \item{territorio}{UF}
  \item{codtroll}{\code{SERCODIGOTROLL}}
}}
\usage{
codterritCONFAZ12
}
\description{
Conjunto de dados contendo o codigo dos territorios utilizado pela CONFAZ.
}
\keyword{datasets}
