% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPISger.R
\name{DEPISwb}
\alias{DEPISwb}
\title{DEPIS - \emph{Webscrapping}}
\usage{
DEPISwb(gerarGen = TRUE)
}
\arguments{
\item{gerarGen}{Logico. Se \code{gerarGen = TRUE}, a planilha \code{GENERICA} e
atualizada no diretorio especifico do \emph{ETL}. O \emph{default} e \code{TRUE}.}
}
\description{
Realiza a raspagem de dados referentes a algumas series do banco DEPIS a partir da
API do SIDRA (\url{https://sidra.ibge.gov.br}).
}
\examples{
#------ Exportando a planilha GENERICA no diretorio.
depis <- DEPISwb()

}
\author{
Luiz Eduardo Gomes, \email{luiz.gomes@ipea.gov.br} ou \email{gomes.leduardo@gmail.com}.
}
