% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\name{discordDatas}
\alias{discordDatas}
\title{Discordancia de datas}
\usage{
discordDatas(serie)
}
\arguments{
\item{serie}{Vetor contendo o \code{SERCODIGOTROLL},\emph{banco}, \emph{periodicidade}
e/ou \emph{status} da(s) serie(s) requisitada(s).}
}
\description{
Retorna a discordancia de datas nas series a partir do \code{SERCODIGOTROLL},
\emph{banco}, \emph{periodicidade} e/ou \emph{status}.
}
\examples{
# ------ Multiplas series
disc <- discordDatas(serie = c("gm12_DOW12","ABATE12_ABPENO12","MTE12_SALMIN12"))

}
\author{
Luiz Eduardo Gomes, \email{luiz.gomes@ipea.gov.br} ou \email{gomes.leduardo@gmail.com}.
}
