% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generica.R
\name{generica}
\alias{generica}
\title{Planilha GENERICA para verificacao}
\usage{
generica(serie)
}
\arguments{
\item{serie}{Vetor contendo o \code{SERCODIGOTROLL}
da(s) serie(s) requisitada(s).}
}
\description{
Retorna a planilha \code{GENERICA} para verificacao a
partir do \code{SERCODIGOTROLL}.
}
\examples{
# ------ Multiplas series, exibindo grafico
gen_verif <- generica(serie = c("GM12_DOW12", "ABATE12_ABPENO12",
"MTE12_SALMIN12"))

}
\author{
Luiz Eduardo Gomes, \email{luiz.gomes@ipea.gov.br}
ou \email{gomes.leduardo@gmail.com}.
}
