% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\docType{data}
\name{lista.interface}
\alias{lista.interface}
\title{Registro da interface de atualizacao das series}
\format{Banco de dados com 13783 observacao(oes) e 6 variavel(is):
\describe{
  \item{SERCODIGO}{Codigo}
  \item{PERNOME}{Periodicidade}
  \item{SERSTATUS}{Status}
  \item{interface}{Interface}
  \item{CodSGS}{Codigo SGS}
  \item{BASNOME}{Nome da base}
}}
\usage{
lista.interface
}
\description{
Registro da interface de atualizacao das series: Generica ou SGS.
}
\keyword{datasets}
