% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\docType{data}
\name{responsavel.series}
\alias{responsavel.series}
\title{Registro dos responsaveis das series do IpeaData (Macro)}
\format{Banco de dados com 11042 observacao(oes) e 2 variavel(is):
\describe{
  \item{Variavel}{Codigo das series}
  \item{Responsavel}{Nome do responsavel}
}}
\usage{
responsavel.series
}
\description{
Conjunto de dados contendo informacoes dos responsaveis das
series do IpeaData (Macro) para registro interno.
}
\keyword{datasets}
