% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\docType{data}
\name{series.automaticas}
\alias{series.automaticas}
\title{Registro dos responsaveis das series do IpeaData (Macro)}
\format{Banco de dados com 900 observacao(oes) e 3 variavel(is):
\describe{
  \item{codtroll}{Codigo}
  \item{tipo}{Tipo de automatizacao}
  \item{funcao}{Funcao que realiza a automacao}
}}
\usage{
series.automaticas
}
\description{
Conjunto de dados contendo informacoes a respeito das
series que estao automatizadas (via R) dentre alguns dos procedimentos
padrao (API, Gera, SQL ou webscrapping)
}
\keyword{datasets}
