% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\docType{data}
\name{serresponsavel}
\alias{serresponsavel}
\title{Registro dos responsaveis das series do IpeaData (Macro)}
\format{Banco de dados com 11042 observacao(oes) e 2 variavel(is):
\describe{
  \item{SERCODIGOTROL}{Codigo das series}
  \item{SERRESPONSAVEL}{Nome do responsavel}
}}
\usage{
serresponsavel
}
\description{
Conjunto de dados contendo informacoes dos responsaveis das
series do IpeaData (Macro) para registro interno.
}
\keyword{datasets}
