% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONFAZ12wb.R
\docType{data}
\name{users.ipeadata}
\alias{users.ipeadata}
\title{Registro dos usuarios IpeaData}
\format{Banco de dados com 10 observacao(oes) e 3 variavel(is):
\describe{
  \item{Usuario}{Login do usuario}
  \item{Nome}{Nome do usuario}
  \item{Email}{E-mail}
}}
\usage{
users.ipeadata
}
\description{
Conjunto de dados contendo informacoes dos membros da equipe para registro interno.
}
\keyword{datasets}
