% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SECEX12FOB12.R
\docType{data}
\name{codpaisesSECEX12FOB12}
\alias{codpaisesSECEX12FOB12}
\title{Codigo dos paises - SECEX MDIC}
\format{um data frame com 252 observacao(oes) e 1 variavel(is):
\describe{
  \item{cod_paises}{Codigo dos paises}
}}
\source{
\url{http://www.mdic.gov.br/arquivos/dwnl_1362158650.pdf}
}
\usage{
codpaisesSECEX12FOB12
}
\description{
Conjunto de dados contendo o codigo dos paises utilizado pela SECEX/MDIC.
}
\keyword{datasets}
