% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\name{dados.faltantes}
\alias{dados.faltantes}
\title{Descontinuidade}
\usage{
dados.faltantes(serie, plotar = TRUE)
}
\arguments{
\item{serie}{Vetor contendo o \code{SERCODIGOTROLL},\emph{banco}, \emph{periodicidade}
e/ou \emph{status} da(s) serie(s) requisitada(s).}

\item{plotar}{Logico. Se \code{plotar = TRUE}, o grafico da(s) serie(s) e(sao) exibidos. 
O \emph{default} e \code{TRUE}.}
}
\description{
Retorna a descontinuidade existente nas series a partir do \code{SERCODIGOTROLL},
\emph{banco}, \emph{periodicidade} e/ou \emph{status}.
}
\note{
Foi limitado a exibicao de ate 5 (cinco) series simultaneamente.
}
\examples{
#------ Multiplas series, exibindo grafico
desc1 <- dados.faltantes(serie = c("gm12_DOW12","ABATE12_ABPENO12","MTE12_SALMIN12"))

#------ Banco, nao exibindo grafico
desc2 <- encontra.serie(serie = c("TRIMESTRAL"),plotar = FALSE) # pode demandar tempo!

#------ Plotando mais de 5 series, retorna erro
erro <- encontra.serie(serie = c("TRIMESTRAL")) # pode demandar tempo!

}
\author{
Luiz Eduardo Gomes, \email{luiz.gomes@ipea.gov.br} ou \email{gomes.leduardo@gmail.com}.
}
