% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\name{encontraSerie}
\alias{encontraSerie}
\title{Encontra series}
\usage{
encontraSerie(serie, plotar = TRUE)
}
\arguments{
\item{serie}{Vetor contendo o \code{SERCODIGOTROLL} da(s) serie(s) requisitada(s).}

\item{plotar}{Logico. Se \code{plotar = TRUE}, o grafico da(s) serie(s) e(sao) exibidos.
O \emph{default} e \code{TRUE}.}
}
\description{
Encontra series a partir do \code{SERCODIGOTROLL},
\emph{banco}, \emph{periodicidade} e/ou \emph{status}.
}
\note{
Foi limitado a exibicao de ate 5 (cinco) series simultaneamente.
}
\examples{
# ------ Serie unica, exibindo grafico
serie1 <- encontraSerie(serie = ("SGS366_CDI"))

# ------ Multiplas series, exibindo grafico
serie2 <- encontraSerie(serie = c("ABATE12_ABQUBO12","ABATE12_ABQUBV12"))

# ------ Multiplas series, nao exibindo grafico
serie3 <- encontraSerie(serie = c("gm12","ABATE12_ABPENO12","MTE12_SALMIN12"),
                         plotar = FALSE)

# ------ Serie nao existente (retorna erro)
erro <- encontraSerie(serie = c("serie que n existe"))

}
\author{
Luiz Eduardo Gomes, \email{luiz.gomes@ipea.gov.br} ou \email{gomes.leduardo@gmail.com}.
}
