% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SECEX12FOB12wb.R
\docType{data}
\name{listadepaises.ZE}
\alias{listadepaises.ZE}
\title{Lista dos paises - ZONA DO EURO - SECEX MDIC}
\format{Banco de dados com 39 observacao(oes) e 21 variavel(is):
\describe{
  \item{"Datas"}{Datas}
  \item{"cod_membro_1"}{Codigo do membro 1}
  \item{"cod_membro_2"}{Codigo do membro 2}
  \item{"cod_membro_3"}{Codigo do membro 3}
  \item{"cod_membro_4"}{Codigo do membro 4}
  \item{"cod_membro_5"}{Codigo do membro 5}
  \item{"cod_membro_6"}{Codigo do membro 6}
  \item{"cod_membro_7"}{Codigo do membro 7}
  \item{"cod_membro_8"}{Codigo do membro 8}
  \item{"cod_membro_9"}{Codigo do membro 9}
  \item{"cod_membro_10"}{Codigo do membro 10}
  \item{"cod_membro_11"}{Codigo do membro 11}
  \item{"cod_membro_12"}{Codigo do membro 12}
  \item{"cod_membro_13"}{Codigo do membro 13}
  \item{"cod_membro_14"}{Codigo do membro 14}
  \item{"cod_membro_15"}{Codigo do membro 15}
  \item{"cod_membro_16"}{Codigo do membro 16}
  \item{"cod_membro_17"}{Codigo do membro 17}
  \item{"cod_membro_18"}{Codigo do membro 18}
  \item{"cod_membro_19"}{Codigo do membro 19}
  \item{"cod_membro_20"}{Codigo do membro 20}
}}
\source{
\url{https://pt.wikipedia.org/wiki/Zona_Euro#Membros}
}
\usage{
listadepaises.ZE
}
\description{
Conjunto de dados contendo os codigos dos paises da Zona do Euro utilizado pela SECEX/MDIC.
}
\keyword{datasets}
