% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\docType{data}
\name{series.automaticas}
\alias{series.automaticas}
\title{Registro das series automaticas}
\format{Banco de dados com 896 observacao(oes) e 3 variavel(is):
\describe{
  \item{codtroll}{Codigo das series}
  \item{tipo}{Tipo do metodo de atualizacao}
  \item{funcao}{Nome da funcao}
}}
\usage{
series.automaticas
}
\description{
Registro das series obtidas de forma automatica via Gera, Webscrapping, SQL ou API.
}
\keyword{datasets}
