% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTEIStroll.R
\name{situavar}
\alias{situavar}
\title{Situacao das series}
\usage{
situavar(serie, exportar = TRUE)
}
\arguments{
\item{serie}{Vetor contendo o \code{SERCODIGOTROLL},\emph{banco}, \emph{periodicidade}
e/ou \emph{status} da(s) serie(s) requisitada(s).}

\item{exportar}{Logico. Se \code{exportar = TRUE}, um relatorio \code{.xls} e exportado para o
diretorio do \code{situavar} (\code{///Srjn3/area_corporativa/Projeto_IPEADATA/Geral/PacoteIpeadataRio/situavar}).
O \emph{default} e \code{TRUE}.}
}
\description{
Retorna a situacao das series com relacao a atraso a partir do \code{SERCODIGOTROLL},
\emph{banco}, \emph{periodicidade} e/ou \emph{status}.
}
\note{
A nomenclatura dada ao arquivo sera \emph{situavarAAAAMMDDhhmmss.xls}.
Quando \code{serie} e unitario, o arquivo e nomeado da forma padrao com a 
adicao do conteudo de \code{serie}.
}
\examples{
#------ Multiplos Bancos e serie
sit1 <- situavar(serie = c("PIMPFN12","GM366","ANDIMA4","CONFAZ12_ICMSSP12"))

#------ Periodicidade e bancos
sit2 <- encontra.serie(serie = c("diaria","IGP12"))
                         
#------ Serie nao existente (retorna erro)
erro <- situavar(serie = c("serie que n existe"))

}
\author{
Luiz Eduardo Gomes, \email{luiz.gomes@ipea.gov.br} ou \email{gomes.leduardo@gmail.com}.
}
